Created thestorygoes
Created description of the project (a bit of a manifesto, a bit of a rant).  Will probably update with something better-formed.

-----

I dunno if I'll ever update this.  I'm half-dazed as I write this; if I knew what drunkenness was, I'd probably point to my current state as a comparative level of inebriation; I've slipped to using semicolons and needlessly complex words, for crying out loud!

Plus, I was never good at committing to projects.  Plus, I have no idea how GitHub actually works.  Plus, I need to learn how to program (and honestly I should be using most of my time doing that).

Oh, well.  The worst thing that can happen is that this dies.